Created Grunt.js based build * provided package.json file * separated qrcode.js into separate files under src/ using common.js format * provided .jshintrc and .jsbeautifyrc files, made sources adhere to rules * Gruntfile.js using jshint, jsbeautifier, browserify and uglify to compile into lib/ directory
