+++ .jsbeautifyrc
@@ -1,0 +1,17 @@
+{
+  "indent_size": 2,
+  "indent_char": " ",
+  "indent_with_tabs": false,
+  "preserve_newlines": true,
+  "max_preserve_newlines": 4,
+  "space_in_paren": false,
+  "jslint_happy": false,
+  "brace_style": "collapse",
+  "keep_array_indentation": false,
+  "keep_function_indentation": false,
+  "eval_code": false,
+  "unescape_strings": false,
+  "break_chained_methods": false,
+  "e4x": false,
+  "wrap_line_length": 0
+}
