+++ src/QRErrorCorrectLevel.js
@@ -1,0 +1,12 @@
+//---------------------------------------------------------------------
+// QRErrorCorrectLevel
+//---------------------------------------------------------------------
+
+var QRErrorCorrectLevel = {
+  L: 1,
+  M: 0,
+  Q: 3,
+  H: 2
+};
+
+module.exports = QRErrorCorrectLevel;
